fix bugs annonce

+++ database/migrations/2023_06_10_184307_create_annonces_table.php
@@ -17,7 +17,7 @@ return new class extends Migration
             $table->foreignId('projet_id')->constrained();
 
             $table->dateTime('date_pub');
-            $table->float('cout')->nullable();
+            $table->double('cout')->nullable();
             $table->foreignId('user_id')->constrained();
             $table->timestamps();
         });
